==> 2-MODELING_RF_HYPEROPT.ipynb <==
import re
from pprint import pprint
from datetime import datetime
from collections import Counter
import pandas as pd
import numpy as np

from sklearn.utils import class_weight
from sklearn.preprocessing import LabelEncoder
from category_encoders import TargetEncoder, CatBoostEncoder
from sklearn.model_selection import cross_val_score, StratifiedKFold, RandomizedSearchCV, train_test_split
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn import metrics
from sklearn.feature_selection import SelectFromModel

import xgboost as xgb
from hyperopt import Trials, STATUS_OK, tpe, hp, fmin

method = "target"
assert method in ["label", "target", "onehot", "catboost"]

features = pd.read_csv(f"../data/processed/{method}/train.csv")
test = pd.read_csv(f"../data/processed/{method}/test.csv")

targets = features["label"].values

features.drop(["label"], inplace=True, axis=1)
print(features.shape, test.shape)

### Feature selection

feature_selection = False

if feature_selection:
    tree = ExtraTreesClassifier(n_estimators=50)
    tree = tree.fit(features, targets)

    selector = SelectFromModel(tree, prefit=True)
    selected_features = features.columns[selector.get_support()]

    features = features[selected_features]
    test = test[selected_features]

    print(features.shape, test.shape)

def objective(space):
    classifier = RandomForestClassifier(class_weight=space["class_weight"],
                                        n_estimators=int(space['n_estimators']),
                                        max_depth=int(space['max_depth']),
                                        max_features=space["max_features"],
                                        min_samples_split=int(space["min_samples_split"]),
                                        min_samples_leaf=int(space["min_samples_leaf"]),
                                        bootstrap=space["bootstrap"],
                                        max_samples=space["max_samples"], 
                                        n_jobs=10)
    #classifier.fit(features, targets)    
    scores = cross_val_score(estimator=classifier,
                             n_jobs=2,
                             X=features, 
                             y=targets, 
                             cv=StratifiedKFold(n_splits=5),
                             scoring="f1_micro"
                            )
    mean_score = scores.mean()
    
    return {
        'loss': 1-mean_score, 
        'status': STATUS_OK 
    }

space = {
    'n_estimators': hp.quniform('n_estimators', 250, 350, 10),
    'class_weight': hp.choice("class_weight", ["balanced", None]),
    'max_depth' : hp.quniform('max_depth', 20, 40, 1),
    'max_features': hp.choice("max_features", ["sqrt", "auto", "log2"]),
    "min_samples_split": hp.quniform("min_samples_split", 2, 30, 2),
    'min_samples_leaf' : hp.quniform("min_samples_leaf", 2, 30, 2),
    'bootstrap': hp.choice("bootstrap", [True, False]),
    'max_samples': hp.quniform("max_samples", 0.5, 0.9, 0.1)
}

trials = Trials()
best = fmin(fn=objective, 
            verbose=5,
            space=space,
            algo=tpe.suggest,
            max_evals=50,
            trials=trials)

best["class_weight"] = ["balanced", None][best["class_weight"]]
best["bootstrap"] = [True, False][best["bootstrap"]]
best["max_features"] = ["sqrt", "auto", "log2"][best["max_features"]]
best["n_estimators"] = int(best["n_estimators"])
best["min_samples_split"] = int(best["min_samples_split"])
best["min_samples_leaf"] = int(best["min_samples_leaf"])
best["max_depth"] = int(best["max_depth"])

print(f"best_params :")
pprint(best)

### Cross val score

model = RandomForestClassifier(**best)
scores = cross_val_score(model, features, targets, scoring="f1_micro", verbose=2, n_jobs=5)
print(f"mean score: {np.mean(scores)}")

### Fit best model 

model = RandomForestClassifier(**best)
model.fit(features, targets)

### Submission 

submission = pd.DataFrame()
submission["Id"] = range(len(test))
submission["label"] = model.predict(test)
submission.to_csv(f"../data/submissions/rf_{method}.csv", index=False)

### Save proba

probas = pd.DataFrame()
probas["Id"] = range(len(test))

probas = pd.concat([probas, pd.DataFrame(model.predict_proba(test))], axis=1)

probas.to_csv(f"../data/submissions/rf_probas_{method}.csv", index=False)

# private leaderboard : 0.59928
