==> 5-ENSEMBLING.ipynb <==
import pandas as pd
import numpy as np

rf = pd.read_csv("../data/submissions/rf_probas.csv")
xgb = pd.read_csv("../data/submissions/xgb_probas.csv")
lgb = pd.read_csv("../data/submissions/lgbm_probas.csv")

for df in [rf, xgb, lgb]:
    df.set_index("Id", inplace=True)

ensemble = (rf + xgb + lgb) / 3

ensemble

submission = pd.DataFrame()
submission["Id"] = range(len(rf))

submission["label"] = ensemble.values.argmax(axis=1)

submission

submission.label.value_counts()

submission.to_csv("../data/submissions/ensemble.csv", index=False)
